refactored member view to make the member home page all the routes and now when they click on the logo in the top left it will direct them to the member home page from wherever they are
