+++ src/components/member/MemberHome.jsx
@@ -6,50 +6,59 @@ import './memberHome.css'
 
 
 export const MemberHome = ({currentUser}) => {
+
     const [allRoutes, setAllRoutes] = useState([])
-    const [displayedRoutes, setDisplayedRoutes] = useState([])
-
-
+    // const [selectedType, setSelectedType] = useState('')
+    // const [selectedGrade, setSelectedGrade] = useState('')
+    // const [selectedStyle, setSelectedStyle] = useState('')
+    // const [selectedSetter, setSelectedSetter] = useState('')
+    // const [displayedRoutes, setDisplayedRoutes] = useState([])
 
     const getAndSetAllRoutes = () => {
         getAllRoutes().then(routesArray => {
             setAllRoutes(routesArray)
-        })
+        }) 
     }
 
     useEffect(() => {
-        getAndSetAllRoutes() 
-    },[])
+        getAndSetAllRoutes()
+    }, [])
 
 
-    // sets up the page so that the users ticked routes show on initial render
-    useEffect(() => {
-        const sentRoutes = allRoutes.filter(route => {
-            return route.ticks.some(tick => tick.userId === currentUser.id)
-        })
-        setDisplayedRoutes(sentRoutes)
-    },[allRoutes])
+    // useEffect(() => {
+    //     let filteredRoutes = allRoutes
+
+    //         if (parseInt(selectedType) > 0) {
+    //             filteredRoutes = filteredRoutes.filter(route => route.typeId === parseInt(selectedType))
+    //         }
+    //         if (parseInt(selectedGrade) > 0) {
+    //             filteredRoutes = filteredRoutes.filter(route => route.gradeId === parseInt(selectedGrade))
+    //         }
+    //         if(parseInt(selectedStyle) > 0) {
+    //             filteredRoutes = filteredRoutes.filter(route => route.styleId === parseInt(selectedStyle))
+    //         }
+    //         if (parseInt(selectedSetter) > 0) {
+    //             filteredRoutes = filteredRoutes.filter(route => route.userId === parseInt(selectedSetter))
+    //         }
+
+    //     setDisplayedRoutes(filteredRoutes)
+
+    // }, [allRoutes, selectedType, selectedGrade, selectedStyle, selectedSetter])
 
 
 
-    return(
-        <div className="container-member-home">
 
-                <h2 className="heading">My Routes</h2>
-                
-                <div className="member-route-filter">
-                    <MemberFilterBar 
-                     allRoutes={allRoutes} 
-                        currentUser={currentUser} 
-                        setDisplayedRoutes={setDisplayedRoutes}
-                    />
-                </div>
-                <div className="routes">
-                    {displayedRoutes.map(route => {
-                        return <MemberRouteObject route={route} key={route.id} />
+    return (
+        <div className="employee-home">
+            <h2 className="heading">Gym Routes</h2>
+            <div className="employee-route-filter">
+
+            </div>
+            <div className="routes">
+                    {allRoutes.map(route => {
+                        return <MemberRouteObject route={route} getAndSetAllRoutes={getAndSetAllRoutes} currentUser={currentUser} key={route.id} allRoutes={allRoutes}/>   
                     })}
-                </div>
-
+            </div>
         </div>
     )
 }
